Add initial version of sidebar for elfeed
The code is based on sidebar-mu4e, and works reasonably okay for feeds.

The feed URLs are displayed directly in the sidebar, and it would be
nice to change them to the feed names. Optionally, it would be nice to
show author names in addition.

Tags are not shown. It may be useful to organize feeds into folders
based on tags.

Aggregated feeds don't work, because we search in elfeed using the
hostname of the entry, but that is different for the post for entries in
aggregated feeds, for example Planet feeds.